pyproject.toml
[project]
name = "5-8-psy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.18.4",
    "datasets>=4.8.4",
    "fastapi>=0.135.2",
    "google-genai>=1.68.0",
    "langchain-google-genai>=4.2.1",
    "langchain-text-splitters>=1.1.1",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pymupdf4llm>=1.27.2.2",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "python-telegram-bot>=22.7",
    "qdrant-client>=1.17.1",
    "ragas>=0.4.3",
    "redis>=7.4.0",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.42.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
]
